requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
telethon==1.33.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django_import_export==3.3.7
